utf8_to_utf16le_avx512.S: implement dependency-chain break in 2-byte fast path
An idea of Daniel, we change the 2-byte fast path to always
process 32 or 33 characters, depending on whether a two byte
sequence straddles bytes 32/33.  This makes it very fast to
compute the start of the next vector of input, permitting
the next iteration of the loop to start much quicker.