fuzz: Give fuzzer independent data per func
Fundamentally a coverage driven fuzzer is an optimization
algorithm, where the goal-state is 100% code coverage. As
it currently stands this fuzzer is trying to optimize it's
input so that is valid UTF-8 AND UTF-16be/le. By giving
each function it's own independent input we allow the fuzzer
to optimize coverage for each function individually.