simplify the conversion fuzzer to get smaller function signatures
by using an array of function pointers to lambdas, the function signatures
are much smaller since they do not include all of the functions.

fuzzing speed increased from about 13 to 18 kHz, mostly because I made
the fuzz case array static so it does not need to be initialized every round.