add .clang-format file
this is based on LLVM.

sorting includes is disabled, since that is significant in this
project for amalgamation and more to work properly.

IndentPPDirectives is set to beforehash. This is in order to make
files with a lot of preprocessor logic more comprehensible.
common_defs.h is the motivating example.

enforce newlines at end: this makes diffs more readable.