add clang format file
this is based on LLVM, taken from the fastfloat project
which also disables sort includes.

in addition to that:

IndentPPDirectives is set to beforehash. This is in order to files with
a lot of preprocessor logic more comprehensible with common_defs.h
being the motivating example.

enforce newlines at end. this makes diffs more readable.