introduce clang format (#563)
* add clang format script

this is intended both for running locally and in CI

* add .clang-format file

this is based on LLVM.

sorting includes is disabled, since that is significant in this
project for amalgamation and more to work properly.

IndentPPDirectives is set to beforehash. This is in order to make
files with a lot of preprocessor logic more comprehensible.
common_defs.h is the motivating example.

enforce newlines at end: this makes diffs more readable.

* exclude benchmark competition files from formatting

this requires clang format 18 or later

the motivation is that those files are imported from upstream and
it makes sense to preserve their original formatting. In case they
are updated it is easier to see what changed.

* preemptively disable problematic formatting in {haswell,icelake}/begin.h

clang-format does not understand the SIMDUTF_DISABLE_GCC_WARNING macro
and treats it as a function call, putting whitespace around the "minus".

* preemptively fix formatting in common_defs.h

clang-format does not understand the SIMDUTF_DISABLE_GCC_WARNING macro
and treats it as a function call, putting whitespace around the "minus".

* format entire code base with clang-format-18

* add docker file for running clang format

the files use my username "pauldreik" since pushing to
simdutf requires more rights than I currently have.

* add script for clang format using docker

this is intended for users who want to use the blessed clang format
version. it is often so that small changes in clang format versions
cause formatting changes.

* add ci job for clang format