base64: factor out common decoding template
This is the first step of larger refactoring. While base64 decoding
itself is not complicated, the correct handling of different options
we provide is quite difficult.

Having this piece of code in a single place would make future
extending the code (like adding a new target) easier and less
error-prone.

The main change is making most free-standing procedures methods
of the `block64` class.