arm64: alternate base64 decode logic to optimize base64url decode (#753)
Adapts the westmere's translation logic for delta_hash by using bits 7:3
instead of bits 7:4 for hi_nibble previously.
New lookup table is designed so that when lo & hi is done, it results in error
bytes being mapped to 0. Spaces are mapped to 1.
Since hi is obtained by TBL of hi_bits 7:4, it results in hi = 0 for all
input bytes > 127 with MSB set.

This makes the base64 decode logic symmetrical for base64 and base64url.
Thus the extra logic to handle '_' is not not required and decode should be
faster for base64url inputs.

Signed-off-by: Shreesh Adiga <16567adigashreesh@gmail.com>